Added Neural Net, Training, & Testing of Model
Created the class for my neural network, currently named VertexCountNet
Based much of my design off of the design of other people’s neural networks, which I commented in links to their tutorials

main.py: Program that trains and saves a convolutional neural network that can count the number of vertices in a graph

testModel.py: Loads and tests how accurate a saved model is against a set of test data.

+++ transforms.py
@@ -91,5 +91,5 @@ class ToTensor(object):
         # numpy image: H x W x C
         # torch image: C x H x W
         image = image.transpose((2, 0, 1))
-        return {'image': torch.from_numpy(image),
+        return {'image': torch.from_numpy(image).double(),
                 'vertices': vertices}
